+++ cuneiform_src/Kern/rstr/src/mmx_addr.c
@@ -67,20 +67,20 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 /**** standart address *****/
 /***************************/
 extern WORD all_scalarf(PWORD,PWORD,WORD);
-extern int all_scalar (PWORD,PWORD);
+extern LONG all_scalar (PWORD,PWORD);
 #ifdef PC_TYPE
 /***************************/
 /****   MMX address    *****/
 /***************************/
-extern int MMX_scalarf(PWORD,PWORD,WORD);
-extern int MMX_scalar (PWORD,PWORD);
+extern LONG MMX_scalarf(PWORD,PWORD,WORD);
+extern LONG MMX_scalar (PWORD,PWORD);
 
 #endif
 /*******************************/
 /**** set standart address *****/
 /*******************************/
 WORD (*scalarf)(PWORD,PWORD,WORD)=all_scalarf;   //AK 14.03.97
-int (*scalar) (PWORD,PWORD)=all_scalar;
+LONG (*scalar) (PWORD,PWORD)=all_scalar;
 
 
 /*******************************/
